requirements.txt
PyQt5
pytesseract
Pillow
openai
google-generativeai
pyinstaller
